requirements.txt
asgiref==3.5.2
black==22.3.0
click==8.1.3
dj-database-url==0.5.0
Django==4.0.5
django-crispy-forms==1.14.0
flake8==4.0.1
gunicorn==20.1.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
sqlparse==0.4.2
tomli==2.0.1
